optimizing code 1
I tried to reduce the number of dictionary calls by saving the results of the dictionary calls during the duplicate_label_check and the VRP_first().  I tried to do the same with the VRP_dominance_check_update(), but something wasn't working correctly (so I hashed out the update).  I'll have to ask Dan about what he thinks is going on.